Simple client-server applications using the sockets API
Develop software to transfer a text file using connection-based byte streams from a server to a client. A command-line argument to the server specifies the file name, and the server must deliver this file to each client that contacts it, possibly sending concurrently to several clients. Each client receives the encoded data and interprets it appropriately to recreate an exact copy of the original file, again using a file name specified on the command line. The file may contain any number of lines and each line is terminated by a single linefeed. The total number of characters, including the linefeed at the end, can be no more than eighty.